fix: throw exception for sniffs not belonging to any standards
If sniffs don't match any standards in the `groupBy` closure in `groupSniffsByStandard` method it should throw an exception instead of returning a default of `null`.

When returning a default of `null`, unknown sniffs would've been grouped as `null`. This is seen as unexpected behaviour because we use the standards to retrieve the sniffs in the first place. So if a sniff is unknown that means that something has gone wrong, so throw an exception.